@@ -377,13 +377,15 @@ public final class AppKitBackend: AppBackend {
             .with(\.colorScheme, isDark ? .dark : .light)
     }
 
-    public func setRootEnvironmentChangeHandler(to action: @escaping () -> Void) {
+    public func setRootEnvironmentChangeHandler(to action: @escaping @Sendable @MainActor () -> Void) {
         DistributedNotificationCenter.default.addObserver(
             forName: .AppleInterfaceThemeChangedNotification,
             object: nil,
             queue: OperationQueue.main
         ) { notification in
-            action()
+            Task { @MainActor in
+                action()
+            }
         }
 
         // This doesn't strictly affect the root environment, but it does require us
@@ -395,7 +397,9 @@ public final class AppKitBackend: AppBackend {
             queue: OperationQueue.main
         ) { notification in
             // Self.scrollBarWidth has changed
-            action()
+            Task { @MainActor in
+                action()
+            }
         }
 
         NotificationCenter.default.addObserver(
@@ -403,7 +407,9 @@ public final class AppKitBackend: AppBackend {
             object: nil,
             queue: .main
         ) { _ in
-            action()
+            Task { @MainActor in
+                action()
+            }
         }
     }
 
@@ -2219,6 +2225,7 @@ enum AssociationPolicy {
 }
 
 // Source: https://gist.github.com/sindresorhus/3580ce9426fff8fafb1677341fca4815
+@MainActor
 final class ObjectAssociation<T: Any> {
     private let policy: AssociationPolicy
 
@@ -2271,6 +2278,7 @@ extension NSControl {
     typealias ActionClosure = ((NSControl) -> Void)
     typealias EditClosure = ((NSTextField) -> Void)
 
+    @MainActor
     struct AssociatedKeys {
         static let onActionClosure = ObjectAssociation<ActionClosure>()
         static let onEditClosure = ObjectAssociation<EditClosure>()
