@@ -2,6 +2,7 @@ import SwiftCrossUI
 import UIKit
 
 /// The context associated with an instance of ``Representable``.
+@MainActor
 public struct UIViewRepresentableContext<Representable: UIViewRepresentable> {
     public let coordinator: Representable.Coordinator
     public internal(set) var environment: EnvironmentValues
@@ -152,7 +153,7 @@ extension View where Self: UIViewRepresentable {
 
         let size = representingWidget.representable.sizeThatFits(
             proposedSize,
-            uiView: representingWidget.subview,
+            uiView: representingWidget.view,
             context: representingWidget.context!
         )
 
@@ -183,27 +184,34 @@ final class ViewRepresentingWidget<Representable: UIViewRepresentable>: BaseView
     var context: Representable.Context?
     var subviewConstraints: [NSLayoutConstraint] = []
 
-    lazy var subview: Representable.UIViewType = {
-        let view = representable.makeUIView(context: context!)
+    var subview: Representable.UIViewType?
 
-        self.addSubview(view)
+    var view: Representable.UIViewType {
+        if let subview {
+            return subview
+        } else {
+            let view = representable.makeUIView(context: context!)
 
-        view.translatesAutoresizingMaskIntoConstraints = false
-        subviewConstraints = [
-            view.topAnchor.constraint(equalTo: self.topAnchor),
-            view.leadingAnchor.constraint(equalTo: self.leadingAnchor),
-            view.trailingAnchor.constraint(equalTo: self.trailingAnchor),
-            view.bottomAnchor.constraint(equalTo: self.bottomAnchor),
-        ]
-        NSLayoutConstraint.activate(subviewConstraints)
+            self.addSubview(view)
 
-        return view
-    }()
+            view.translatesAutoresizingMaskIntoConstraints = false
+            subviewConstraints = [
+                view.topAnchor.constraint(equalTo: self.topAnchor),
+                view.leadingAnchor.constraint(equalTo: self.leadingAnchor),
+                view.trailingAnchor.constraint(equalTo: self.trailingAnchor),
+                view.bottomAnchor.constraint(equalTo: self.bottomAnchor),
+            ]
+            NSLayoutConstraint.activate(subviewConstraints)
+
+            subview = view
+            return view
+        }
+    }
 
     func update(with environment: EnvironmentValues) {
         if var context {
             context.environment = environment
-            representable.updateUIView(subview, context: context)
+            representable.updateUIView(view, context: context)
             self.context = context
         } else {
             let context = Representable.Context(
@@ -211,7 +219,7 @@ final class ViewRepresentingWidget<Representable: UIViewRepresentable>: BaseView
                 environment: environment
             )
             self.context = context
-            representable.updateUIView(subview, context: context)
+            representable.updateUIView(view, context: context)
         }
     }
 
@@ -221,8 +229,10 @@ final class ViewRepresentingWidget<Representable: UIViewRepresentable>: BaseView
     }
 
     deinit {
-        if let context {
-            Representable.dismantleUIView(subview, coordinator: context.coordinator)
+        if let context, let subview {
+            Task { @MainActor in
+                Representable.dismantleUIView(subview, coordinator: context.coordinator)
+            }
         }
     }
 }
